Decompile DRA CheckDarkMetamorphosisInput
