+++ config/symbols.hd.dra.txt
@@ -507,7 +507,7 @@ func_80110394 = 0x801100D8;
 CheckGravityBootsInput = 0x80110214;
 CheckQuarterCircleForwardInput = 0x801103E8;
 CheckBackForwardInput = 0x80110560;
-func_80110968 = 0x801106AC;
+CheckDarkMetamorphosisInput = 0x801106AC;
 func_80110BC8 = 0x8011090C;
 func_80110DF8 = 0x80110B3C;
 func_80111018 = 0x80110D5C;
