+++ config/symbols.us.dra.txt
@@ -950,6 +950,7 @@ g_Cd_overlayLastBlockSize = 0x80137F90;
 g_VabId = 0x80137F94;
 g_WasFacingLeft = 0x80137FC0;
 g_WasFacingLeft2 = 0x80137FC4;
+g_WasFacingLeft3 = 0x80137FC8;
 g_WingSmashButtonCounter = 0x80137FF4;
 g_WingSmashButtonTimer = 0x80137FF8;
 g_WingSmashTimer = 0x80137FFC;
