+++ src/dra/6D59C.c
@@ -1370,173 +1370,3 @@ s16 func_80110394(void) {
     return 0;
 }
 
-bool CheckGravityBootsInput(void) {
-    switch (g_ButtonCombo[COMBO_GRAVITY_BOOTS].buttonsCorrect) {
-    case 0:
-        if ((g_Player.padTapped & PAD_DOWN) && (g_Player.padHeld == 0)) {
-            g_ButtonCombo[COMBO_GRAVITY_BOOTS].timer = 16;
-            g_ButtonCombo[COMBO_GRAVITY_BOOTS].buttonsCorrect++;
-        }
-        break;
-    case 1:
-        if (g_Player.padTapped & PAD_UP) {
-            g_ButtonCombo[COMBO_GRAVITY_BOOTS].timer = 16;
-            g_ButtonCombo[COMBO_GRAVITY_BOOTS].buttonsCorrect++;
-        } else {
-            if (--g_ButtonCombo[COMBO_GRAVITY_BOOTS].timer == 0) {
-                g_ButtonCombo[COMBO_GRAVITY_BOOTS].buttonsCorrect = 0;
-            }
-        }
-        break;
-    case 2:
-        if ((g_ButtonCombo[COMBO_GRAVITY_BOOTS].timer != 0) &&
-            --g_ButtonCombo[COMBO_GRAVITY_BOOTS].timer == 0) {
-            g_ButtonCombo[COMBO_GRAVITY_BOOTS].buttonsCorrect = 0;
-            break;
-        }
-        if (IsRelicActive(RELIC_GRAVITY_BOOTS) &&
-            (g_Player.padTapped & PAD_CROSS) && !(g_Player.unk46 & 0x8000) &&
-            ((PLAYER.step == 2) ||
-             ((PLAYER.step == 4) && (g_Player.unk44 & 1)))) {
-            if (g_Player.unk72 == 0) {
-                if (func_800FF064(1) >= 0) {
-                    func_8010E9A4();
-                    g_ButtonCombo[COMBO_GRAVITY_BOOTS].buttonsCorrect = 0;
-                    return 1;
-                }
-            } else {
-                g_ButtonCombo[COMBO_GRAVITY_BOOTS].buttonsCorrect = 0;
-            }
-        }
-        break;
-    }
-    return 0;
-}
-
-// Used for wolf charge, and for weapons which have a move using this pattern
-bool CheckQuarterCircleForwardInput(void) {
-    s32 directionsPressed;
-    s32 down_forward;
-    s32 forward;
-
-    directionsPressed =
-        g_Player.padPressed & (PAD_UP | PAD_RIGHT | PAD_DOWN | PAD_LEFT);
-    if (!g_WasFacingLeft) {
-        forward = PAD_RIGHT;
-    } else {
-        forward = PAD_LEFT;
-    }
-    if (!g_WasFacingLeft) {
-        down_forward = PAD_DOWN + PAD_RIGHT;
-    } else {
-        down_forward = PAD_DOWN + PAD_LEFT;
-    }
-    switch (g_ButtonCombo[COMBO_QCF].buttonsCorrect) {
-    case 0:
-        g_WasFacingLeft = PLAYER.facingLeft;
-        if (directionsPressed == PAD_DOWN) {
-            g_ButtonCombo[COMBO_QCF].timer = 20;
-            g_ButtonCombo[COMBO_QCF].buttonsCorrect++;
-            break;
-        }
-        if (--g_ButtonCombo[COMBO_QCF].timer == 0) {
-            g_ButtonCombo[COMBO_QCF].buttonsCorrect = 0;
-        }
-        break;
-    case 1:
-        if (directionsPressed == down_forward) {
-            g_ButtonCombo[COMBO_QCF].timer = 20;
-            g_ButtonCombo[COMBO_QCF].buttonsCorrect++;
-            break;
-        }
-        if (--g_ButtonCombo[COMBO_QCF].timer == 0) {
-            g_ButtonCombo[COMBO_QCF].buttonsCorrect = 0;
-        }
-        break;
-    case 2:
-        if (directionsPressed == forward) {
-            g_ButtonCombo[COMBO_QCF].timer = 20;
-            g_ButtonCombo[COMBO_QCF].buttonsCorrect = COMBO_COMPLETE;
-            break;
-        }
-        if (--g_ButtonCombo[COMBO_QCF].timer == 0) {
-            g_ButtonCombo[COMBO_QCF].buttonsCorrect = 0;
-        }
-        break;
-    default:
-        if (g_ButtonCombo[COMBO_QCF].timer != 0) {
-            if (g_Player.unk72 != 0) {
-                g_ButtonCombo[COMBO_QCF].buttonsCorrect = 0;
-                break;
-            }
-            if (--g_ButtonCombo[1].timer == 0) {
-                g_ButtonCombo[COMBO_QCF].buttonsCorrect = 0;
-                g_WasFacingLeft = PLAYER.facingLeft;
-            }
-        }
-    }
-    return 0;
-}
-
-bool CheckBackForwardInput(void) {
-    s32 directionsPressed;
-    s32 backward;
-    s32 forward;
-
-    directionsPressed =
-        g_Player.padPressed & (PAD_UP | PAD_RIGHT | PAD_DOWN | PAD_LEFT);
-    if (!g_WasFacingLeft2) {
-        forward = PAD_RIGHT;
-    } else {
-        forward = PAD_LEFT;
-    }
-    switch (g_ButtonCombo[COMBO_BF].buttonsCorrect) {
-    case 0:
-        g_WasFacingLeft2 = PLAYER.facingLeft;
-        if (!g_WasFacingLeft2) {
-            backward = PAD_LEFT;
-        } else {
-            backward = PAD_RIGHT;
-        }
-        if (directionsPressed == backward) {
-            g_ButtonCombo[COMBO_BF].timer = 15;
-            g_ButtonCombo[COMBO_BF].buttonsCorrect++;
-        } else if (--g_ButtonCombo[COMBO_BF].timer == 0) {
-            g_ButtonCombo[COMBO_BF].buttonsCorrect = 0;
-        }
-        break;
-    case 1:
-        if (directionsPressed == forward) {
-            g_ButtonCombo[COMBO_BF].timer = 15;
-            g_ButtonCombo[COMBO_BF].buttonsCorrect = COMBO_COMPLETE;
-            break;
-        } else if (--g_ButtonCombo[COMBO_BF].timer == 0) {
-            g_ButtonCombo[COMBO_BF].buttonsCorrect = 0;
-        }
-        break;
-    case 2:
-    default:
-        if (g_Player.unk72 != 0) {
-            g_ButtonCombo[COMBO_BF].buttonsCorrect = 0;
-            break;
-        }
-        if (g_ButtonCombo[COMBO_BF].timer != 0 &&
-            --g_ButtonCombo[COMBO_BF].timer == 0) {
-            g_ButtonCombo[COMBO_BF].buttonsCorrect = 0;
-            g_WasFacingLeft2 = PLAYER.facingLeft;
-        }
-    }
-    return 0;
-}
-
-INCLUDE_ASM("dra/nonmatchings/6D59C", func_80110968);
-
-INCLUDE_ASM("dra/nonmatchings/6D59C", func_80110BC8);
-
-INCLUDE_ASM("dra/nonmatchings/6D59C", func_80110DF8);
-
-INCLUDE_ASM("dra/nonmatchings/6D59C", func_80111018);
-
-INCLUDE_ASM("dra/nonmatchings/6D59C", func_801112AC);
-
-INCLUDE_ASM("dra/nonmatchings/6D59C", func_8011151C);
